refactor:  input, output 기능을 Console 객체로 통합
