--- src/utils/consolePrint.js
@@ -1,8 +1,0 @@
-function consolePrint(query) {
-  if (arguments.length !== 1) {
-    throw new Error('arguments must be 1');
-  }
-  console.log(query);
-}
-
-export default consolePrint;
